Cell Initialization
Created a more modular way of initializing the cell objects. Object types are stored in the objectTypes array and the cell objects will iterate through this array and set a property matching each type's name and then will set it to null. We use null instead of undefined for better error catching. If undefined is thrown, then we know there is an error somewhere.